Rename Dash ent rypoint to dashboard_app.py and update docs
