+++ scripts/profile_dash.py
@@ -1,4 +1,4 @@
-from main import profile_startup
+from dashboard_app import profile_startup
 
 if __name__ == "__main__":
     profile_startup()
